PrivateRoute now also uses props.children - Refactoring PrivateRoute - Fixing imports for tests - Fixing some lint errors
