+++ src/components/common/PrivateRoute.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Redirect, Route } from 'react-router-dom'
 
-const PrivateRoute = ({ render: Component, redirectPath, condition, ...rest }) => {
+const PrivateRoute = ({ render: Component, redirectPath, condition, children, ...rest }) => {
   return (
-    <Route {...rest} render={props =>
+    <Route {...rest} render={(props) =>
       condition
-        ? (<Component {...props} />)
+        ? (Component && <Component {...props} />) || children
         : (
           <Redirect
             to={{
